Finalized implementation of `dash::Atomic` by removing `AtomicAddress`. This also includes many bugfixes, unit tests and cleanup
