+++ dash/include/dash/Shared.h
@@ -7,7 +7,7 @@
 #include <dash/GlobRef.h>
 #include <dash/Allocator.h>
 
-#include <dash/atomic/AtomicAddress.h>
+#include <dash/atomic/GlobAtomicRef.h>
 
 #include <dash/iterator/GlobIter.h>
 
@@ -38,7 +38,7 @@ public:
   typedef       GlobPtr<value_type>                     pointer;
   typedef const GlobPtr<value_type>               const_pointer;
 
-  typedef dash::AtomicAddress<ElementType>          atomic_type;
+  typedef GlobRef<Atomic<ElementType>>          atomic_ref_type;
 
 private:
   typedef dash::GlobMem<
@@ -50,7 +50,7 @@ private:
   friend void swap(Shared<T_> & a, Shared<T_> & b);
 
 public:
-  atomic_type atomic;
+  atomic_ref_type atomic;
 
 public:
   /**
@@ -81,7 +81,9 @@ public:
       ds.dtype,
       _owner,
       _team->dart_id());
-    atomic = atomic_type(_ptr.dart_gptr());
+    atomic._set_dart_gptr(_ptr.dart_gptr());
+    // ensure that atomic proxy is initialized on all units
+    team.barrier();
     DASH_LOG_DEBUG_VAR("Shared.Shared(team,owner) >", _ptr);
   }
